Split out WeightInfo into weight_info.rs
So that we could use a script to generate the weights later.
